pyproject.toml
[tool.poetry]
name = "livi"
version = "0.0.1"
description = "Lyrics Informed Version Identification Framework."
authors = ["Research <research@deezer.com>"]
readme = "README.md"
repository = "https://github.com/deezer/version-identification-lyrics"
homepage = "https://github.com/deezer/version-identification-lyrics"

packages = [
  { include = "livi", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
torch = "^2.6.0"
torchaudio = "^2.6.0"
torchvision = "*"
transformers = "^4.34.0"
einops = "0.6.*"
datasets = "^3.6.0"
deezer-datasource = "^4.2.5"
sentence-transformers = "4.1.0"
soundfile = "^0.13.1"
librosa = "^0.11.0"
hydra-core = "^1.3.2"
wandb = "^0.20.1"
loguru = "^0.7.3"
webdataset = "^1.0.2"
instrumentalvocal = "^1.3.0"
pydantic-settings = "^2.10.1"
pathlib = "^1.0.1"
pre-commit = "^4.3.0"
typer = "^0.17.3"
lightning = "^2.5.4"
nnaudio = "^0.3.3"
timm = "^1.0.19"
omegaconf = "^2.3.0"
matplotlib = "^3.10.6"
torchcodec = "^0.6.0"
langchain-text-splitters = "^0.3.11"
bm25s = "^0.2.13"
flagembedding = "^1.3.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E741", "E742", "E743", "F", "I"]
ignore = ["T201", "T203"]

[[tool.poetry.source]]
name = "research"
url = "https://artifacts.deez.re/repository/python-research/simple/"
priority = "primary"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.poetry.scripts]
livi-audio-encoder = "livi.apps.audio_encoder.cli:app"
livi-frozen-encoder = "livi.apps.frozen_encoder.cli:app"
livi-retrieval-eval = "livi.apps.retrieval_eval.cli:app"
livi-audio-baselines = "livi.apps.audio_baselines.cli:app"
livi-data = "livi.core.data.cli:app"